Add connection with Gui and server - bad messages are sent

+++ client/src/main/java/playing/FieldPosition.java
@@ -2,13 +2,12 @@ package playing;
 
 public class FieldPosition {
 
-
     private final Integer x;
     private final Integer y;
 
-    FieldPosition(double x, double y) {
-        this.x = (int) x;
-        this.y = (int) y;
+    FieldPosition(int x, int y) {
+        this.x = x;
+        this.y = y;
     }
 
     public Integer getX() {
